Moves boards into its own module. DRY for success
